Added project settings from extensions
