+++ src/extension-host/extension-host.ts
@@ -13,6 +13,7 @@ import { startProjectLookupService } from '@extension-host/services/project-look
 import { registerCommand } from '@shared/services/command.service';
 import { initialize as initializeMenuData } from '@extension-host/services/menu-data.service-host';
 import { initialize as initializeSettingsService } from '@extension-host/services/settings.service-host';
+import { startProjectSettingsService } from '@extension-host/services/project-settings.service-host';
 
 // #region Test logs
 
@@ -65,6 +66,7 @@ networkService
     // Prepare for contributions to be contributed from extensions before the extensions come online
     await initializeMenuData();
     await initializeSettingsService();
+    await startProjectSettingsService();
 
     // The extension service locks down importing other modules, so be careful what runs after it
     await extensionService.initialize();
